IS-722 modified rendering description with values for domain cards
