+++ app/decorators/daggerheart_character/stats_decorator.rb
@@ -62,22 +62,6 @@ module DaggerheartCharacter
         stance_bonus
     end
 
-    def armor_score # rubocop: disable Metrics/AbcSize
-      @armor_score ||=
-        base_armor_score +
-        item_bonuses.pluck('armor_score').compact.sum +
-        sum(bonuses.pluck('armor_score')) +
-        sum(dynamic_bonuses.pluck('armor_score')) +
-        sum(static_feat_bonuses.pluck('armor_score')) +
-        sum(dynamic_feat_bonuses.pluck('armor_score')) +
-        sum(static_item_bonuses.pluck('armor_score')) +
-        sum(dynamic_item_bonuses.pluck('armor_score'))
-    end
-
-    def armor_slots
-      @armor_slots ||= armor_score
-    end
-
     def health_max # rubocop: disable Metrics/AbcSize
       @health_max ||=
         data.health_max +
