+++ app/serializers/feat_serializer.rb
@@ -10,10 +10,11 @@ class FeatSerializer < ApplicationSerializer
   end
 
   def description
-    Charkeeper::Container.resolve('markdown').call(
+    result = Charkeeper::Container.resolve('markdown').call(
       value: object.description[I18n.locale.to_s],
       version: (context ? (context[:version] || nil) : nil),
       initial_version: '0.3.20'
     )
+    context && context[:gsub] ? result&.gsub(/{{[a-z]+}}/, 'x') : result
   end
 end
